refactor(kv-store)!: move IndexedDB backend to deprecated/ and rename its export
The IndexedDB backend now lives at @aztec/kv-store/deprecated/indexeddb (was
@aztec/kv-store/indexeddb, now removed) and its exports are marked @deprecated.
New browser code should use @aztec/kv-store/browser (SQLite-OPFS).
