+++ yarn-project/kv-store/package.json
@@ -7,7 +7,7 @@
     "./interfaces": "./dest/interfaces/index.js",
     "./lmdb": "./dest/lmdb/index.js",
     "./lmdb-v2": "./dest/lmdb-v2/index.js",
-    "./indexeddb": "./dest/indexeddb/index.js",
+    "./deprecated/indexeddb": "./dest/deprecated/indexeddb/index.js",
     "./browser": "./dest/browser/index.js",
     "./sqlite-opfs": "./dest/sqlite-opfs/index.js",
     "./stores": "./dest/stores/index.js"
