+++ yarn-project/kv-store/src/deprecated/indexeddb/array.test.ts
@@ -1,5 +1,5 @@
-import { describeAztecArray } from '../interfaces/array_test_suite.js';
-import { mockLogger } from '../interfaces/utils.js';
+import { describeAztecArray } from '../../interfaces/array_test_suite.js';
+import { mockLogger } from '../../interfaces/utils.js';
 import { AztecIndexedDBStore } from './store.js';
 
 describe('IndexedDBArray', () => {
