+++ yarn-project/kv-store/src/deprecated/indexeddb/map.test.ts
@@ -1,5 +1,5 @@
-import { describeAztecMap } from '../interfaces/map_test_suite.js';
-import { mockLogger } from '../interfaces/utils.js';
+import { describeAztecMap } from '../../interfaces/map_test_suite.js';
+import { mockLogger } from '../../interfaces/utils.js';
 import { AztecIndexedDBStore } from './store.js';
 
 describe('IndexedDBMap', () => {
