+++ yarn-project/kv-store/src/deprecated/indexeddb/multi_map.test.ts
@@ -1,5 +1,5 @@
-import { describeAztecMultiMap } from '../interfaces/multi_map_test_suite.js';
-import { mockLogger } from '../interfaces/utils.js';
+import { describeAztecMultiMap } from '../../interfaces/multi_map_test_suite.js';
+import { mockLogger } from '../../interfaces/utils.js';
 import { AztecIndexedDBStore } from './store.js';
 
 describe('IndexedDBMap', () => {
