+++ yarn-project/kv-store/src/deprecated/indexeddb/set.test.ts
@@ -1,5 +1,5 @@
-import { describeAztecSet } from '../interfaces/set_test_suite.js';
-import { mockLogger } from '../interfaces/utils.js';
+import { describeAztecSet } from '../../interfaces/set_test_suite.js';
+import { mockLogger } from '../../interfaces/utils.js';
 import { AztecIndexedDBStore } from './store.js';
 
 describe('IndexedDBSet', () => {
