+++ yarn-project/kv-store/src/deprecated/indexeddb/set.ts
@@ -1,7 +1,7 @@
 import type { IDBPDatabase, IDBPObjectStore } from 'idb';
 
-import type { Key, Range } from '../interfaces/common.js';
-import type { AztecAsyncSet } from '../interfaces/set.js';
+import type { Key, Range } from '../../interfaces/common.js';
+import type { AztecAsyncSet } from '../../interfaces/set.js';
 import { IndexedDBAztecMap } from './map.js';
 import type { AztecIDBSchema } from './store.js';
